+++ Sources/SwiftAvroCore/Codable/Decodable/AvroBinaryDecoder.swift
@@ -1,3 +1,21 @@
+//
+//  AvroBinaryDecoder.swift
+//
+//  Created by Kacper Kawecki on 24/03/21.
+//  Copyright © 2021 by Kacper Kawecki and the project authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import Foundation
 
 internal final class AvroBinaryDecoder: Decoder {
